Hoist the navigator-swap lifecycle onto the shared base
